Resolved a merge conflict.
Rebase. This combines the following commits:

feat(BitwiseNavbar): Started to update the notification indicator so
that it is consistent with the mockups (placing the number in a bubble
to the top right of the bell)

refactor(BitwiseNavbar): Refactored the previous approach; Now using the actual notification count; Set the cap to 9 based on the limited space that is available in the nav bar

feat(BitwiseNavbar): The count bubble will only appear when the number of notifications is more than zero.

feat(BitwiseNavbar): The NotificationButtonContainer now supports both the mobile and desktop perspectives.

fix(BitwiseNavbar): Fixed a responsiveness issue with the notification counter at the intersection of the mobile/desktop perspectives; Made it so that there would always be a pointer for the NotificationButtonContainer

fix(various): No longer hard-coding the colors for the notification counter; Made the counter look better when looking at it from the mobile perspective